Command History Input Parser
Now you can write (or copy/paste) commands to the input parser.
This can be used to ‘load’ images you have drawn. Just copy the output
from the Command History to a text-file.
This is of course not meant as a final load/save solution - but it’s
better than nothing.